Added initial Camera Influence Volumes
[ADDED] Camera Influencer Volumes:
This is a script that can be added on any game object, requiring a Camera Controller to be present on the player object, and the player object having a camera object as child. When the player object gets close to one or more influence volumes, it will lerp the camera location between the volumes based on a weight value.

[CHANGED] Camera controller
Camera controller has been overhauled. When there are no nearby influence volumes, it will act like the old camera controller.

[KNOWN ISSUES]
Lerping between 3 influence volumes is a bit buggy still.

[OTHER CHANGES] Cleaned up variable ordering and naming / commenting